updated slides services with more access checking

+++ app/services/slides/delete_slide_service.rb
@@ -2,13 +2,16 @@
 
 module Slides
   class DeleteSlideService < BaseService
-    attr_reader :slide_params
+    attr_reader :current_user, :slide_params
 
-    def initialize(slide_params)
+    def initialize(current_user, slide_params)
+      @current_user = current_user
       @slide_params = slide_params
     end
 
     def call
+      playlist = Playlist.find(slide_params[:playlist_id])
+      authorize playlist, :use?
       slide = Slide.find_by!(slide_params)
 
       slide.destroy
